pyproject.toml
[project]
name = "fastapi-agents"
version = "0.1.0"
description = "A FastAPI extension for integrating common AI agent frameworks."
authors = [
    { name = "Blair Hudson", email = "blairhudson@users.noreply.github.com" },
]
urls = { "Repository" = "https://github.com/blairhudson/fastapi-agents" }
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = [
    "fastapi",
    "ai",
    "agents",
    "smolagents",
    "llama-index",
    "pydantic-ai",
]

requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
]

[build-system]
requires = ["hatchling==1.26.3"]
build-backend = "hatchling.build"

[project.optional-dependencies]
llama-index = [
    "llama-index>=0.12.10",
]
pydantic-ai = [
    "pydantic-ai>=0.0.18",
]
smolagents = [
    "smolagents>=1.2.2",
    "litellm>=1.57.8",
]
uvicorn = [
    "uvicorn>=0.34.0",
]
security = [
    "python-multipart>=0.0.20",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.4",
]
